Backend/requirements.txt
Django>=5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.1
requests>=2.31.0
google-generativeai>=0.3.2
python-dotenv>=1.0.1
gunicorn>=21.2.0
dj-database-url>=2.1.0
